fix: don't use TS in specs
